mmc: k1: map the register width, not the pointer size
spacemit_sdhci_set_aib_mmc1_io() maps three single 32-bit registers
(accessed with readl/writel) but passed sizeof(uintptr_t) as the
mapping length, which is 8 bytes on a 64-bit build. Map sizeof(u32),
the actual register width.

Fixes a low-severity review nit. Squash into "mmc: k1: add sdhci
platform driver" before upstreaming.

+++ drivers/mmc/spacemit_sdhci.c
@@ -303,11 +303,11 @@ static void spacemit_sdhci_set_aib_mmc1_io(struct sdhci_host *host, int voltage)
 		return;
 
 	aib_mmc1_io = map_sysmem((uintptr_t)plat->aib_mmc1_io_reg,
-				 sizeof(uintptr_t));
+				 sizeof(u32));
 	apbc_asfar = map_sysmem((uintptr_t)plat->apbc_asfar_reg,
-				sizeof(uintptr_t));
+				sizeof(u32));
 	apbc_assar = map_sysmem((uintptr_t)plat->apbc_assar_reg,
-				sizeof(uintptr_t));
+				sizeof(u32));
 
 	writel(AKEY_ASFAR, apbc_asfar);
 	writel(AKEY_ASSAR, apbc_assar);
